FINALLY fixed “new best fit slow to appear in top 10” (details below).
Symptom was that a new Individual would be created by crossover, evaluate to a new best fitness, but then not appear in the logged “population top 10” for hundreds of steps.

Bug was that all participants in the TournamentGroup would have their fitness (MOF and scalar) evaluated if not already cached. This did NOT happen to the new offspring. This was hidden because the per-flock-simulation logging that happens in GP appeared the same in both cases.

I assume the delay was waiting for the new offspring to be randomly selected for its first tournament, at which time it would get the MOF/scalar fitness caching.

I fixed it by adding TournamentGroup::custom_eval normally nullptr. The MOF version of Population::evolutionStep() stores a lambda there to do its intended MOF/scalar fitness evaluation.