OK I was mistaken yesterday, but I zeroed in on where the tree corruption happens.
Added special tracking macro "check_Individual_13()" and used it to verify that the Individual's trees are OK until the end of creating the population and scanning over all of them.
But the first call to after the population build finds the corruption. Before it ever gets to mof_eval() in Population::evolutionStep().
Changed GP::run_gp_flock_simulation() to select single threaded operation for GP_not_GA.
Added pop_position and duplicate_tree to Individual.