OK I fixed the recently introduced error running in “Boid::GP_not_GA=false” mode.
I had removed a “individual->treeValue();” to get the GP case (Boid::GP_not_GA=true) to work, but that broke the GA case (Boid::GP_not_GA=false) and presumably all previous FunctionSets.

Apparently I need to make that optional, default ON but have a way to set it to OFF for the Boid::GP_not_GA=true case.