Revert back to per chunk avoidance score for GP.
For GP, in GP::multiObjectiveFitnessOfFlock(), revert back from flock.get_per_boid_avoid_score(), to avoid, that is, require ALL boids to be on the correct side of ALL obstacles.

Also tried 4X run like that.

Also tried experiment with removing Neighbor_1_Velocity and Neighbor_1_Offset for FunctionSet to see if that made them concentrate on obstacle avoidance. It did not. Reverted the change.