Tried running GA version, but with all FlockParameters set to the hand-tuned numbers.
Noticed that the behavior was STILL bad: flying through obstacles.
I was going to disable GP::evoflock_ga_crossover() but when I made it print, it looked like it was never called.
Tried reducing population size to more quickly get to the point where crossovers were used, that made all in the population have zero fitness, causing the rest of the run to be skipped.