WIP restrict annotation vectors to selected boid and its neighbors.
The first part was easy but it was more complicated and more computation to know if a given boid is a neighbor of the selected boid. This version works but I plan to reimplement it with a better approach.

Added:
Boid::selected() a bool meaning “I am selected”.
Boid::selectedBoid()
Boid::amINeighborOfSelectedBoid()

Changed Flock::updateSelectedBoid()