Broke off Obstacle::doesAgentViolateConstraint()
By adjusting margins around obstacle surface, I realized that we cannot wait for the agent to be (eg) outside the big sphere, because when it is, there is no longer a collision predicted. That messes with the logic of doesAgentViolateConstraint(), enforceConstraint(), and “test_obs_constraint()”

I guess we need to test each boid against each obstacle regardless of prediction, or do it as now but predict where the agent will be on the following sim step?