Think I convinced myself that Obstacle::enforceConstraint() works for the ExcludeFrom::neither case.
Improved logging and traps based on selected Boid.

With Obstacle avodiance turned off, and I have seen no violation of the PlaneObstacle. As expected I see selected boid collide with  PlaneObstacle, but it is always pushed back to the correct side.

Running a long duration test for piece of mind.