Adjusting braking weight for predictive avodiance.
Measuring by count of obstacle collisions per 2000 step run for a fixed obstacle set.

For the sphere-and-plane set, increasing from 0.2 to 0.4 to 0.6 kept getting better (fewer collisions).

For the sphere-and-little-spheres set, increasing from 0.2 was best (fewer collisions).

Not clear what setting was best.

(Also made sure to init boid speed to zero and set previous_position to init position, but neither moved the needle. of colision count)