Trying to reduce number of obs collisions.
Found a bug in steer_for_predictive_avoidance() then needed to do lots of rebalancing.

Did a lot of fiddling with the way predictive aND static avoidance get blended in steer_to_avoid().

In GP::evoflock_ga_function_set_handmade() changed min_time_to_collide BACK to 0.8 from yesterday’s 0.4.

Tried to get it to notice a collision about to happen, then select that boid and pause, but that kept making it slow and non-responsive. Could not find the problem, gave up on that side quest.