Fix problem when slow Draw frames seemed to accumlate.
Changes to classes AnimationTimer and RollingAverage in util.

Lower threshold (from 2x to 1.1x) of average recent frame durations at which history is reset to flush accumulating slowness, probably from mouse gestures.

Changed from using std::setprecision(5) -- which is "sticky" -- to std::format("{:.5}") -- whose effect is local.  (This was the source of yesterday's bug getting 0.03 when should be 0.03333)

Add <format> to Utilities.h.