Miscellaneous WIP: slow frame after mouse moves,unnormalized separationScore().
In drawAnnotationForBoidAndNeighbors() lines to neighbors in gray of separation score.

Count and log each invocation of setMouseMoveCallback().

Try increasing the threshold above frameDurationTarget() to reset history in AnimationClock.

Code to trap separationScore() bigger than 1. I first saw this bug yesterday, several times, then as I was trying to catch it in the act, it stopped happening.