Proposed fix for recent sim slowness.
After the commit “WIP change boid neighbors to emphasize boids in front.” The time for a fitness test increased from about 0.31 sec to 0.4 sec.

It turned out to be an essentially unrelated edit, causing the Boid's list of 200 flockmate pointers to be copied each sim step.

Turning that off brought the fitness test time down close to 0.255 sec.

Early sim steps run 3 fitness tests, so the times for those in the log were: 0.94, 0.12, 0.77