Got it to run GA mode without calling run_flock_simulation() inside GpFunc "Run_Flock"
It had been returning the MOF of a given GpTree.

Next step will be to make it instead return the FlockParameters for that GpTree, to be used in run_ga_gp_flock_simulation(). It currently comes from fp_from_ga_individual(individual) called at the top of run_ga_gp_flock_simulation().