Remove “can we do an eval of a const tree?” experiment.
I now think the time to copy a tree is is low, and a fraction of the time to eval.

Adds complexity without benefit to EvoFlock code.

Code is 14 months old “TODO 20240628 can we do an eval of a const tree?”

Leaving in LP for now, moving bits from EvoFlock to graveyard.

Aside: there is nothing wrong with the idea of LP providing a way to evaluate a const GpTree. However, especially using steady state evolution (SSGA/SSGP), we also want to cache the result of evaluating an Individual. (Although, caching an Individual’s fitness is not the same as caching a tree’s value. Maybe GpTree::eval() should be const, and the caching be only in Individual::set/getFitness()?