Still investigating GP misbehavior.
Thought I was missing a lock/mutex for setting log_flock. But adding it did not fix the issue.

Decided to verify that log_flock is in fact one of the four Flocks created in GP::run_flock_simulation(). It was NOT. Worse, it is the same Flock for all evo steps!!