WIP on "too few boid steps" in GP mode.
Put in an assert to detect the frame count mismatch, but still not obvious where the problem is.

Change Boid::setSpeedAfterObstacleCollision() to just bounce at full speed. Hides the symptom where they seem to end up in one octant of big sphere.

Clean up ordering of defs in EvoFlock.h.