Init population to meet both tree size constraints and “sensor API preservation”
Add GP::evoflockGpValidateTree() to test for presence of full sensor API in a tree.

Add FS::getValidateTreeFunction() and FS::setValidateTreeFunction()

In EF::runOneFlockEvolution(), set the validate tree function to GP::evoflockGpValidateTree().

In FS::newMakeRandomTree, extend size_ok() to include a call to the validate tree function.

In Population::evolutionStep(TournamentGroup ranked_group...) add a retry loop looking for a crossover which preserves calls to full sensor API. But this frequently fails (as many as 1000 times in a row) especially early in the run. So some individuals will not have full sensor API. I was thinking of scaling down the fitness of these to avoid taking over the population. But by 20-30% of the evo run, these seem to have disappeared.