Progress on evoflockGpValidateTree() not working during evo pop init.
Added logging to GP::evoflockGpValidateTree(), then removed it later.

The problem seemed to be related to copying the FS in EF::runOneFlockEvolution(). went away when I refactored to use a reference to the original GP FS defined as a static global in GP.h. I do not understand the failure mode. Need more experiments.

Put assert in GP::evoflockGpValidateTree() to trap when it got a reference different from the one selected in GP::evoflockGpValidateTree(). Changed measureUsageSensorAPI() and run_flock_simulation() to force them ti use the correct copy of FS.